+++ latch/types/file.py
@@ -77,9 +77,11 @@ class LatchFile(FlyteFile):
         self._path_generated = False
 
         if is_valid_url(self.path) and remote_path is None:
+            self._raw_remote_path = str(path)
             self._remote_path = str(path)
         else:
             self._remote_path = None if remote_path is None else str(remote_path)
+            self._raw_remote_path = self._remote_path
 
         if kwargs.get("downloader") is not None:
             super().__init__(self.path, kwargs["downloader"], self._remote_path)