+++ latch_cli/main.py
@@ -192,7 +192,15 @@ def init(
     type=bool,
     help="Generate a Dockerfile with arguments needed for Snakemake compatability",
 )
-def dockerfile(pkg_root: str, snakemake: bool = False):
+@click.option(
+    "-n",
+    "--nextflow",
+    is_flag=True,
+    default=False,
+    type=bool,
+    help="Generate a Dockerfile with arguments needed for Nextflow compatability",
+)
+def dockerfile(pkg_root: str, snakemake: bool = False, nextflow: bool = False):
     """Generates a user editable dockerfile for a workflow and saves under `pkg_root/Dockerfile`.
 
     Visit docs.latch.bio to learn more.
@@ -201,24 +209,37 @@ def dockerfile(pkg_root: str, snakemake: bool = False):
     crash_handler.message = "Failed to generate Dockerfile."
     crash_handler.pkg_root = pkg_root
 
-    from latch_cli.docker_utils import generate_dockerfile
+    if snakemake is True and nextflow is True:
+        click.secho(
+            f"Please specify at most one workflow type to generate metadata for. Use"
+            f" either `--snakemake` or `--nextflow`.",
+            fg="red",
+        )
+        raise click.exceptions.Exit(1)
 
-    source = Path(pkg_root)
-    dest = source / "Dockerfile"
-    if dest.exists() and not click.confirm(
-        f"Dockerfile already exists at `{dest}`. Overwrite?"
-    ):
-        return
+    from latch_cli.docker_utils import generate_dockerfile, generate_dockerignore
+
     workflow_type = WorkflowType.latchbiosdk
     if snakemake is True:
         workflow_type = WorkflowType.snakemake
-    generate_dockerfile(source, dest, wf_type=workflow_type)
+    elif nextflow is True:
+        workflow_type = WorkflowType.nextflow
 
-    click.secho(f"Successfully generated dockerfile `{dest}`", fg="green")
+    source = Path(pkg_root)
+    generate_dockerfile(source, wf_type=workflow_type)
+
+    if not click.confirm(f"Generate a .dockerignore?"):
+        return
+    generate_dockerignore(source, wf_type=workflow_type)
 
 
 @main.command("generate-metadata")
-@click.argument("config_file", nargs=1, type=click.Path(exists=True, path_type=Path))
+@click.argument(
+    "config_file",
+    required=False,
+    nargs=1,
+    type=click.Path(exists=True, path_type=Path, dir_okay=False),
+)
 @click.option(
     "--yes",
     "-y",
@@ -229,11 +250,30 @@ def dockerfile(pkg_root: str, snakemake: bool = False):
     ),
 )
 @click.option(
+    "--snakemake",
+    "-s",
+    is_flag=True,
+    default=False,
+    type=bool,
+    help="Generate Latch metadata for Snakemake.",
+)
+@click.option(
+    "--nextflow",
+    "-n",
+    is_flag=True,
+    default=False,
+    type=bool,
+    help="Generate Latch metadata for Nextflow.",
+)
+@click.option(
     "--no-infer-files",
     "-I",
     is_flag=True,
     default=False,
-    help="Don't parse strings with common file extensions as file parameters.",
+    help=(
+        "Don't parse strings with common file extensions as file parameters. Only"
+        " supported for Snakemake workflows."
+    ),
 )
 @click.option(
     "--no-defaults",
@@ -243,17 +283,107 @@ def dockerfile(pkg_root: str, snakemake: bool = False):
     help="Don't generate defaults for parameters.",
 )
 def generate_metadata(
-    config_file: Path, yes: bool, no_infer_files: bool, no_defaults: bool
+    config_file: Optional[Path],
+    snakemake: bool,
+    nextflow: bool,
+    yes: bool,
+    no_infer_files: bool,
+    no_defaults: bool,
 ):
-    """Generate a `latch_metadata.py` file from a Snakemake config file"""
+    """Generate a `latch_metadata.py` file from a config file"""
 
-    from latch_cli.snakemake.config.parser import generate_metadata
+    if snakemake is True and nextflow is True:
+        click.secho(
+            f"Please specify only one workflow type to generate metadata for. Use"
+            f" either `--snakemake` or `--nextflow`.",
+            fg="red",
+        )
+        raise click.exceptions.Exit(1)
 
-    generate_metadata(
-        config_file,
-        skip_confirmation=yes,
-        infer_files=not no_infer_files,
-        generate_defaults=not no_defaults,
+    if nextflow is True:
+        from latch_cli.nextflow.config import generate_metadata
+
+        if config_file is None:
+            config_file = Path("nextflow_schema.json")
+
+        generate_metadata(
+            config_file,
+            skip_confirmation=yes,
+            generate_defaults=not no_defaults,
+        )
+    else:
+
+        from latch_cli.snakemake.config.parser import generate_metadata
+
+        if config_file is None:
+            click.secho(
+                dedent("""
+                Please provide a config file for Snakemake workflows:
+                `latch generate-metadata <config_file_path> --snakemake`
+                """),
+                fg="red",
+            )
+            raise click.exceptions.Exit(1)
+
+        generate_metadata(
+            config_file,
+            skip_confirmation=yes,
+            infer_files=not no_infer_files,
+            generate_defaults=not no_defaults,
+        )
+
+
+@main.command("generate-entrypoint")
+@click.argument("pkg_root", nargs=1, type=click.Path(exists=True, path_type=Path))
+@click.option(
+    "--nf-script",
+    type=click.Path(exists=True, dir_okay=False, path_type=Path),
+    help="Path to the nextflow entrypoint to register.",
+)
+@click.option(
+    "--execution-profile",
+    type=str,
+    default=None,
+    help="Set execution profile for Nextflow workflow",
+)
+def generate_entrypoint(
+    pkg_root: Path, nf_script: Path, execution_profile: Optional[str]
+):
+    """Generate a `wf/entrypoint.py` file from a Nextflow workflow"""
+
+    import latch.types.metadata as metadata
+    from latch_cli.nextflow.workflow import generate_nextflow_workflow
+    from latch_cli.services.register.utils import import_module_by_path
+
+    dest = pkg_root / "wf" / "entrypoint.py"
+    dest.parent.mkdir(exist_ok=True)
+
+    if dest.exists() and not click.confirm(
+        f"Nextflow entrypoint already exists at `{dest}`. Overwrite?"
+    ):
+        return
+
+    meta = pkg_root / "latch_metadata" / "__init__.py"
+    if meta.exists():
+        click.echo(f"Using metadata file {click.style(meta, italic=True)}")
+        import_module_by_path(meta)
+
+    if metadata._nextflow_metadata is None:
+        click.secho(
+            dedent("""
+            Failed to generate Nextflow entrypoint.
+            Make sure the project root contains a `latch_metadata/__init__.py`
+            with a `NextflowMetadata` object defined.
+            """),
+            fg="red",
+        )
+        raise click.exceptions.Exit(1)
+
+    generate_nextflow_workflow(
+        pkg_root,
+        nf_script,
+        dest,
+        execution_profile=execution_profile,
     )
 
 
@@ -412,6 +542,25 @@ def execute(
         " provided."
     ),
 )
+@click.option(
+    "--nf-script",
+    type=click.Path(exists=True, dir_okay=False, path_type=Path),
+    default=None,
+    help="Path to a nextflow script to register.",
+)
+@click.option(
+    "--redownload-dependencies",
+    type=bool,
+    is_flag=True,
+    default=False,
+    help="Redownload external Nextflow dependencies",
+)
+@click.option(
+    "--execution-profile",
+    type=str,
+    default=None,
+    help="Set execution profile for Nextflow workflow",
+)
 @requires_login
 def register(
     pkg_root: str,
@@ -422,6 +571,9 @@ def register(
     open: bool,
     snakefile: Optional[Path],
     cache_tasks: bool,
+    nf_script: Optional[Path],
+    redownload_dependencies: bool,
+    execution_profile: Optional[str],
 ):
     """Register local workflow code to Latch.
 
@@ -442,6 +594,9 @@ def register(
         skip_confirmation=yes,
         open=open,
         snakefile=snakefile,
+        nf_script=nf_script,
+        nf_redownload_dependencies=redownload_dependencies,
+        nf_execution_profile=execution_profile,
         progress_plain=(docker_progress == "auto" and not sys.stdout.isatty())
         or docker_progress == "plain",
         use_new_centromere=use_new_centromere,