+++ CHANGELOG.md
@@ -16,6 +16,25 @@ Types of changes
 
 # Latch SDK Changelog
 
+## 2.45.6 - 2024-06-04
+
+### Added
+
+* Nextflow
+  + Support for generating metadata from `nextflow_schema.json`
+  + Command to generate Nextflow entrypoint without registering
+
+### Fixed
+
++ Make `LatchRule` class a dataclass
+
+## 2.45.5 - 2024-06-03
+
+### Added
+
++ Support for uploading Nextflow workflows
+
+
 ## 2.45.4 - 2024-05-29
 
 ### Changed